+++ xibalba-proto/src/error.rs
@@ -92,7 +92,7 @@ pub enum UrlError {
     InvalidByte(usize),
 }
 
-#[derive(Debug, Clone, PartialEq, Eq)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
 #[non_exhaustive]
 pub enum ParseError {
     /// Response is incomplete; need more data.